Moved selector code to the formSelector file
