+++ app/client/src/selectors/entitiesSelector.ts
@@ -16,10 +16,7 @@ import ImageAlt from "assets/images/placeholder-image.svg";
 import { CanvasWidgetsReduxState } from "../reducers/entityReducers/canvasWidgetsReducer";
 import { MAIN_CONTAINER_WIDGET_ID } from "constants/WidgetConstants";
 import { AppStoreState } from "reducers/entityReducers/appReducer";
-import {
-  GenerateCRUDEnabledPluginMap,
-  UIComponentTypes,
-} from "../api/PluginApi";
+import { GenerateCRUDEnabledPluginMap } from "../api/PluginApi";
 import { APP_MODE } from "entities/App";
 
 export const getEntities = (state: AppState): AppState["entities"] =>
@@ -303,24 +300,6 @@ export const getAction = (
 ): Action | undefined => {
   const action = find(state.entities.actions, (a) => a.config.id === actionId);
   return action ? action.config : undefined;
-};
-
-export const getUIComponent = (
-  state: AppState,
-  pluginId: string,
-  allPlugins: any,
-) => {
-  // Adding uiComponent field to switch form type to UQI or allow for backward compatibility
-  const plugin = allPlugins.find((plugin: any) =>
-    !!pluginId ? plugin.id === pluginId : false,
-  );
-  // Defaults to old value, new value can be DBEditorForm or UQIDBEditorForm
-  let uiComponent = UIComponentTypes.DbEditorForm;
-  if (plugin) {
-    uiComponent = plugin.uiComponent;
-  }
-
-  return uiComponent;
 };
 
 export function getCurrentPageNameByActionId(
