+++ app/client/src/selectors/formSelectors.ts
@@ -1,6 +1,7 @@
 import { getFormValues, isValid, getFormInitialValues } from "redux-form";
 import { AppState } from "reducers";
 import { ActionData } from "reducers/entityReducers/actionsReducer";
+import { UIComponentTypes } from "../api/PluginApi";
 
 type GetFormData = (
   state: AppState,
@@ -22,3 +23,24 @@ export const getApiName = (state: AppState, id: string) => {
 
 export const getFormEvaluationState = (state: AppState) =>
   state.evaluations.formEvaluation;
+
+// Selector to get UIComponent type from the redux state
+export const getUIComponent = (
+  state: AppState,
+  pluginId: string,
+  allPlugins: any,
+) => {
+  let uiComponent = UIComponentTypes.DbEditorForm;
+
+  if (!!pluginId) {
+    // Adding uiComponent field to switch form type to UQI or allow for backward compatibility
+    const plugin = allPlugins.find((plugin: any) =>
+      !!pluginId ? plugin.id === pluginId : false,
+    );
+    // Defaults to old value, new value can be DBEditorForm or UQIDBEditorForm
+    if (plugin) {
+      uiComponent = plugin.uiComponent;
+    }
+  }
+  return uiComponent;
+};
